Deduplicate arch/emscripten/bits/alltypes.h{,in} with include/libc/alltypes.h.in
And fix pthread_attr_t __u.__s type to match musl.

+++ system/lib/pthread/library_pthread.c
@@ -53,7 +53,7 @@ int emscripten_pthread_attr_gettransferredcanvases(const pthread_attr_t* a, cons
 }
 
 int emscripten_pthread_attr_settransferredcanvases(pthread_attr_t* a, const char* str) {
-  a->_a_transferredcanvases = (int)str;
+  a->_a_transferredcanvases = (unsigned long)str;
   return 0;
 }
